Extract edge reporting to optimize_reports
