+++ freqtrade/optimize/optimize_reports.py
@@ -107,3 +107,29 @@ def generate_text_table_strategy(stake_currency: str, max_open_trades: str,
     # Ignore type as floatfmt does allow tuples but mypy does not know that
     return tabulate(tabular_data, headers=headers,
                     floatfmt=floatfmt, tablefmt="pipe")  # type: ignore
+
+
+def generate_edge_table(results: dict) -> str:
+
+    floatfmt = ('s', '.10g', '.2f', '.2f', '.2f', '.2f', 'd', '.d')
+    tabular_data = []
+    headers = ['pair', 'stoploss', 'win rate', 'risk reward ratio',
+                'required risk reward', 'expectancy', 'total number of trades',
+                'average duration (min)']
+
+    for result in results.items():
+        if result[1].nb_trades > 0:
+            tabular_data.append([
+                result[0],
+                result[1].stoploss,
+                result[1].winrate,
+                result[1].risk_reward_ratio,
+                result[1].required_risk_reward,
+                result[1].expectancy,
+                result[1].nb_trades,
+                round(result[1].avg_trade_duration)
+            ])
+
+    # Ignore type as floatfmt does allow tuples but mypy does not know that
+    return tabulate(tabular_data, headers=headers,
+                    floatfmt=floatfmt, tablefmt="pipe")  # type: ignore
